+++ src/CMDBObjectTypes.php
@@ -92,7 +92,7 @@ class CMDBObjectTypes extends Request {
     /**
      * Fetches information about an object type by its title (which could be a "language constant")
      *
-     * @param $title
+     * @param string $title Object title
      *
      * @return array Associative array
      *